quadtree docs
